ILogPanelViewModel.Initialize VisualLineElementGenerator collection
