+++ Reporter/ViewModel.cs
@@ -16,6 +16,7 @@ using Common.Models;
 using Common.UserInterface.Commands;
 using Common.UserInterface.ViewModels;
 using ICSharpCode.AvalonEdit;
+using ICSharpCode.AvalonEdit.Rendering;
 using Microsoft.WindowsAPICodePack.Dialogs;
 using Reporter.Models;
 using Reporter.Utilities;
@@ -171,7 +172,7 @@ namespace Reporter
 
         public async Task Initialize(TextEditor reportText)
         {
-            LogPanelViewModel.Initialize(reportText);
+            LogPanelViewModel.Initialize(reportText, new List<VisualLineElementGenerator>());
 
             await InitializeReportsList();
         }
